fix : fix the model long and in and del the value from entity

+++ mohaymen-codestar-Team02/Models/EdgeEAV/EdgeAttribute.cs
@@ -5,9 +5,9 @@ namespace mohaymen_codestar_Team02.Models.EdgeEAV;
 
 public class EdgeAttribute
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
     [ForeignKey("EntityId")] public virtual EdgeEntity EdgeEntity { get; set; }
     public virtual ICollection<EdgeValue> Values { get; set; } = new List<EdgeValue>();
 }
