+++ mohaymen-codestar-Team02/Models/VertexEAV/VertexAttribute.cs
@@ -6,9 +6,9 @@ namespace mohaymen_codestar_Team02.Models.VertexEAV;
 
 public class VertexAttribute
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
     [ForeignKey("EntityId")] public virtual EdgeEntity EdgeEntity { get; set; } 
     public virtual ICollection<EdgeValue> Values { get; set; } = new List<EdgeValue>();
 }
